+++ examples/basic.rs
@@ -1,5 +1,15 @@
 extern crate uni_app;
 
+#[cfg(target_arch = "wasm32")]
+use wasm_bindgen::prelude::*;
+
+#[cfg(target_arch = "wasm32")]
+#[wasm_bindgen(start)]
+pub fn main_js() -> Result<(), JsValue> {
+    main();
+    Ok(())
+}
+
 fn main() {
     // create the game window (native) or canvas (web)
     let app = uni_app::App::new(uni_app::AppConfig {
@@ -17,7 +27,7 @@ fn main() {
         for evt in app.events.borrow().iter() {
             // print on stdout (native) or js console (web)
             uni_app::App::print(format!("{:?}\n", evt));
-            // exit on key ou mouse press
+            // exit on key or mouse press
             match evt {
                 &uni_app::AppEvent::KeyUp(_) => {
                     uni_app::App::exit();